refactored *_drawers.cpp
